aotf: separate vuetify components from form generation * FormInput should be a configurable interface supporting different compoents * For now just move the config to another file and stick it into the component's data
